% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package on-load hook}
\usage{
.onLoad(libname, pkg)
}
\arguments{
\item{libname}{Library path.}

\item{pkg}{Package name.}
}
\description{
Package on-load hook
}
\keyword{internal}
