% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters-advanced.R
\name{detrendSignal}
\alias{detrendSignal}
\title{Detrend signal}
\usage{
detrendSignal(x, type = c("linear", "constant"), output_assay = "detrended")
}
\arguments{
\item{x}{A \code{PhysioExperiment} object.}

\item{type}{Type of detrending: "linear" or "constant" (mean removal).}

\item{output_assay}{Name for the output assay. Default is "detrended".}
}
\value{
A \code{PhysioExperiment} object with a new assay named \code{output_assay}
containing detrended data. For "constant" type, the channel mean is
subtracted. For "linear" type, a least-squares linear fit is removed.
}
\description{
Removes linear or polynomial trends from the signal.
}
\references{
Oppenheim, A.V. & Willsky, A.S. (1997). "Signals and Systems."
2nd ed. Prentice Hall.
}
\seealso{
\code{\link[=detrendSignals]{detrendSignals()}} for the alternative detrending implementation
with polynomial support, \code{\link[=butterworthFilter]{butterworthFilter()}} for highpass filtering
as an alternative to detrending.
}
